app/requirements.txt

paho-mqtt
flask
flask-cors
